shall running with all existing command

+++ include/fcntl.h
@@ -8,6 +8,6 @@
 struct linux_dirent {
     unsigned long d_ino;
     unsigned long d_off;
-    unsigned long d_reclen;
+    unsigned short d_reclen;
     char          d_name[];
 };
